Moved huey to a new location
