--- app/config_email.py
@@ -1,2 +1,0 @@
-from huey import RedisHuey
-huey = RedisHuey()
