+++ app/email/models.py
@@ -1,0 +1,9 @@
+"""
+filename: models.py
+description: Models for email.
+created by: Omar De La Hoz (oed7416@rit.edu)
+created on: 09/27/18
+"""
+
+from huey import RedisHuey
+huey = RedisHuey()
